requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.2.25
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.62.1
frozenlist==1.8.0
greenlet==3.3.2
idna==3.11
kiwisolver==1.5.0
magic-filter==1.0.12
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.48
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
yarl==1.23.0
